Fold CostAndUsageReportLineItem into LineItem.
