+++ src/test/com/netflix/ice/processor/LineItemTest.java
@@ -26,7 +26,7 @@ import org.junit.Test;
 
 import com.amazonaws.services.s3.model.S3ObjectSummary;
 
-public class CostAndUsageReportLineItemTest {
+public class LineItemTest {
     private static final String resourcesDir = "src/test/resources";
 
 	String[] item = {
@@ -61,7 +61,7 @@ public class CostAndUsageReportLineItemTest {
 		S3ObjectSummary s3ObjectSummary = new S3ObjectSummary();
 		s3ObjectSummary.setLastModified(new Date());
 		CostAndUsageReport caur = new CostAndUsageReport(s3ObjectSummary, new File(resourcesDir, "LineItemTest-Manifest.json"), null, "");
-		LineItem li = new CostAndUsageReportLineItem(false, null, caur);		
+		LineItem li = new LineItem(false, null, caur);		
 		li.setItems(item);
 		assertEquals("Usage quantity is incorrect", Double.parseDouble(li.getUsageQuantity()), 1.0, 0.001);
 	}
@@ -71,7 +71,7 @@ public class CostAndUsageReportLineItemTest {
 		S3ObjectSummary s3ObjectSummary = new S3ObjectSummary();
 		s3ObjectSummary.setLastModified(new Date());
 		CostAndUsageReport caur = new CostAndUsageReport(s3ObjectSummary, new File(resourcesDir, "LineItemTest-Manifest.json"), null, "");
-		LineItem li = new CostAndUsageReportLineItem(false, null, caur);		
+		LineItem li = new LineItem(false, null, caur);		
 		li.setItems(item);
 		assertEquals("Wrong resource tags size", 1, li.getResourceTagsSize());
 		String[] tagsHeader = li.getResourceTagsHeader();
